Initial plug-in architecture.
