+++ fetch-validator-status/Dockerfile
@@ -1,8 +1,8 @@
 FROM bcgovimages/von-image:next-1
 
-RUN pip install pynacl
+RUN pip install pynacl gspread oauth2client
 
-ADD fetch_status.py .
 ADD networks.json .
+ADD *.py ./
 
 ENTRYPOINT ["bash", "-c", "python fetch_status.py $@", "--"]
